test: include the skillpoint tests in UpgradeTests
