+++ RailwayCo/Assets/Tests/EditModeTests/GameLogicTests/TrainAttributeTests.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 using NUnit.Framework;
