+++ RailwayCo/Assets/Tests/EditModeTests/GameLogicTests/UpgradeTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ea66e83af9e3b8c47af92b83c84a20e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
